+++ packages/react/src/components/Box.tsx
@@ -3,7 +3,7 @@ import { ComponentProps, ElementType } from 'react'
 import { styled } from '../styles'
 
 export const Box = styled('div', {
-  padding: '$4',
+  padding: '$6', // 48px
   borderRadius: '$md',
   backgroundColor: '$gray800',
   border: '1px solid $gray600',