+++ packages/react/src/components/TextInput/styles.ts
@@ -7,7 +7,18 @@ export const TextInputContainer = styled('div', {
   boxSizing: 'border-box',
   border: '2px solid $gray900',
   display: 'flex',
-  alignItems: 'baseline',
+  alignItems: 'center',
+
+  variant: {
+    size: {
+      sm: {
+        padding: '$2 $3',
+      },
+      md: {
+        padding: '$3 $4',
+      },
+    },
+  },
 
   '&:has(input:focus)': {
     borderColor: '$ignite300',
@@ -41,8 +52,8 @@ export const Input = styled('input', {
   '&:disabled': {
     cursor: 'not-allowed',
   },
-  
-  '&:placeholder': {
+
+  '&::placeholder': {
     color: '$gray400',
   },
 })