+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dsoldera-ignite-design-system/react
 
+## 2.1.0
+
+### Minor Changes
+
+- update react styles
+
 ## 2.0.0
 
 ### Major Changes